🚚 move metrics to its own mode, remove unused metrics imports
