+++ src/admin/mod.rs
@@ -3,3 +3,4 @@ mod error;
 pub mod listener;
 pub mod liveready;
 mod methods;
+pub mod metrics;
